Move post processor out of DocumentBuilder
