+++ src/docfx/SubCommands/DocumentBuilderWrapper.cs
@@ -7,18 +7,17 @@ namespace Microsoft.DocAsCode.SubCommands
     using System.Collections.Generic;
     using System.Collections.Immutable;
     using System.IO;
-    using System.Linq;
-    using System.Reflection;
     using System.Runtime.Remoting.Lifetime;
+    using System.Reflection;
 
     using Microsoft.DocAsCode;
+    using Microsoft.DocAsCode.Build.Common;
     using Microsoft.DocAsCode.Build.ConceptualDocuments;
-    using Microsoft.DocAsCode.Build.Engine;
     using Microsoft.DocAsCode.Build.ManagedReference;
     using Microsoft.DocAsCode.Build.ResourceFiles;
     using Microsoft.DocAsCode.Build.RestApi;
-    using Microsoft.DocAsCode.Build.Common;
     using Microsoft.DocAsCode.Build.TableOfContents;
+    using Microsoft.DocAsCode.Build.Engine;
     using Microsoft.DocAsCode.Common;
     using Microsoft.DocAsCode.Exceptions;
     using Microsoft.DocAsCode.Plugins;
@@ -91,11 +90,13 @@ namespace Microsoft.DocAsCode.SubCommands
 
         public static void BuildDocument(BuildJsonConfig config, TemplateManager templateManager, string baseDirectory, string outputDirectory, string pluginDirectory)
         {
-            using (var builder = new DocumentBuilder(LoadPluginAssemblies(pluginDirectory)))
+            var assebmlies = LoadPluginAssemblies(pluginDirectory);
+            using (var builder = new DocumentBuilder(config.PostProcessors.ToImmutableArray(), config.GlobalMetadata.ToImmutableDictionary(), assebmlies))
             {
                 builder.IntermediateFolder = config.IntermediateFolder;
                 using (new PerformanceScope("building documents", LogLevel.Info))
                 {
+                    var manifests = new List<Manifest>();
                     foreach (var parameters in ConfigToParameter(config, templateManager, baseDirectory, outputDirectory))
                     {
                         if (parameters.Files.Count == 0)
@@ -105,13 +106,20 @@ namespace Microsoft.DocAsCode.SubCommands
                                 : $"No files found, nothing is generated in version \"{parameters.VersionName}\".");
                             return;
                         }
+                        builder.PrepareMetadata(parameters);
                         if (!string.IsNullOrEmpty(parameters.VersionName))
                         {
                             Logger.LogInfo($"Start building for version: {parameters.VersionName}");
                         }
-                        builder.Build(parameters);
+                        manifests.Add(builder.Build(parameters));
                     }
-                    builder.SaveManifest(outputDirectory);
+                    var generatedManifest = TemplateProcessor.MergeManifest(manifests);
+
+                    builder.RemoveDuplicateOutputFiles(generatedManifest.Files);
+                    builder.PostProcess(generatedManifest, outputDirectory);
+
+                    // Save to manifest.json & .manifest(deprecated)
+                    TemplateProcessor.SaveManifest(generatedManifest, outputDirectory);
                 }
             }
         }
