+++ src/docfx/SubCommands/MetadataMerger.cs
@@ -59,7 +59,7 @@ namespace Microsoft.DocAsCode.SubCommands
             };
             var fc = new FileCollection(parameters.Files);
             fc.RemoveAll(x => "toc.yml".Equals(Path.GetFileName(x.File), StringComparison.OrdinalIgnoreCase));
-            var models = DocumentBuilder.Build(
+            var models = DocumentBuilderCore.Build(
                 p,
                 new DocumentBuildParameters
                 {
